refactor order controller to endpoints refactor order operations add testing

+++ src/DemoShop.Application/Features/Order/DTOs/OrderItemResponse.cs
@@ -6,6 +6,6 @@ public sealed record OrderItemResponse
     public required string ProductName { get; init; }
     public required string ProductThumbnail { get; init; }
     public required int Quantity { get; init; }
-    public required int UnitPrice { get; init; }
-    public required int TotalPrice { get; init; }
+    public required decimal UnitPrice { get; init; }
+    public required decimal TotalPrice { get; init; }
 }
