+++ src/DemoShop.Application/Features/Order/DTOs/OrderResponse.cs
@@ -1,4 +1,8 @@
+#region
+
 using DemoShop.Domain.Order.Enums;
+
+#endregion
 
 namespace DemoShop.Application.Features.Order.DTOs;
 
@@ -7,7 +11,7 @@ public sealed record OrderResponse
     public required int Id { get; init; }
     public required int UserId { get; init; }
     public required IReadOnlyCollection<OrderItemResponse> Items { get; init; }
-    public required int Amount { get; init; }
+    public required decimal Amount { get; init; }
     public required OrderStatus Status { get; init; }
     public required DateTime Created { get; init; }
 }
